# xtask coverage-libs using criterion to better assess performance