--- a/xtask/Cargo.toml
+++ b/xtask/Cargo.toml
@@ -34,6 +34,8 @@ timing = "0.2.3"
 ansi_rgb = "0.2.0"
 dhat = { version = "0.2.4", optional = true }
 humansize = "1.1.1"
+itertools = "0.10.3"
+criterion = "0.3.5"
 
 [features]
 dhat-on = ["dhat"]