@MichaReiser 2022-01-06
> > ```
> >     [jquery.min.js] Total Time: 27.4061ms (tokenization: 4.2022ms, parsing: 13.2653ms, tree_sink: 9.9386ms)        
> > ```
> 
> Do we still need to print that summary line?
> 
> That's cool. Finally a more reliable way to benchmark our parser :)

I haven't seen this explanation. Seems yes, it is useful.  It just feels a bit out of place, maybe it looks better if it's printed as a summary at the very end, without indention and after the `Diagnostics`?
>  The "Total Time" line is interesting to run with the oneliners below and gather a summary.