@xunilrj 2022-01-06
> ... maybe it looks better if it's printed as a summary at the very end, without indention and after the `Diagnostics`?

Done